Add files via upload
we have 3 KPIs which list the number of calls in total, the average customer satisfaction score, and the average call duration in minutes. Below them, we have a line chart to show the trend of calls during October(that’s the only month available, the dataset is information about calls made in October 2020), a doughnut chart to show through which channel the call was made. Then we have a map, a column chart for the reason the person called, and last but not least, a bar chart indicating the response time. My favorite part was the slicers that we can see on the left side. We can use them to filter our data and thus make our dashboard interactive.